refactor: 전체 코드를 chess 패키지 하위에 두고, Pawn과 PawnTest는 pieces 패키지 하위에 둔다
