+++ src/main/java/chess/pieces/Pawn.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.pieces;
 
 public class Pawn {
     String color;
