+++ src/test/java/chess/pieces/PawnTest.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.pieces;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
